Move BombData to dedicated BombData.obj
