+++ config/ghidra_ns_to_obj.csv
@@ -1,6 +1,7 @@
 AnmManager,th06::AnmManager
 AnmVm,th06::AnmVm
 AsciiManager,th06::AsciiManager,th06::StageMenu
+BombData,th06::BombData
 BulletManager,th06::BulletManager
 Chain,th06::Chain,th06::ChainElem
 EclManager,th06::EclManager
