+++ src/diffbuild.hpp
@@ -32,6 +32,7 @@
 
 #ifdef DIFFBUILD
 #define DIFFABLE_EXTERN(type, name) extern "C" type name;
+#define DIFFABLE_EXTERN_ARRAY(type, size, name) extern "C" type name[size];
 #define DIFFABLE_STATIC(type, name) extern "C" type name;
 #define DIFFABLE_STATIC_ARRAY(type, size, name) extern "C" type name[size];
 // This macro is meant to be used like so:
@@ -54,6 +55,7 @@
     template <> type DIFFBUILD_HIDE_NAME_##name[size]
 #else
 #define DIFFABLE_EXTERN(type, name) extern type name;
+#define DIFFABLE_EXTERN_ARRAY(type, size, name) extern "C" type name[size];
 #define DIFFABLE_STATIC(type, name) type name;
 #define DIFFABLE_STATIC_ARRAY(type, size, name) type name[size];
 #define DIFFABLE_STATIC_ASSIGN(type, name) type name
